+++ wpull/options.py
@@ -965,6 +965,13 @@ class AppArgumentParser(argparse.ArgumentParser):
             default=True,
             help=_('don’t ignore filters for robots.txt'),
         )
+        group.add_argument(
+            '--no-strong-redirects',
+            dest='strong_redirects',
+            action='store_false',
+            default=True,
+            help=_('don’t implicitly allow span hosts for redirects'),
+        )
 
     def _add_phantomjs_args(self):
         group = self.add_argument_group(_('PhantomJS'))