+++ doc/changelog.rst
@@ -5,7 +5,9 @@ What's New
 0.XX (2013-03-DD)
 ==================
 
-* robots.txt is now unconditionally fetched. Use ``--no-strong-robots`` to disable this behavior.
+* ``robots.txt`` is now unconditionally fetched. Use ``--no-strong-robots`` to disable this behavior.
+* Redirects implicitly span hosts. Use ``--no-strong-redirects`` to disable this behavior.
+* Scripting: ``should_fetch()`` info dict now contains ``reason`` as a key.
 
 
 0.23 (2013-03-07)